The SpringBox layout is now usable as a base class.
The old GraphStream SpringBox graph layout is now a cut into a
BarnesHutLayout class that serves as base for several force-based layout
algorithms and a new layout.implementations package contains a SpringBox
class that inherit BarnesHutLayout.

Nothing changed in the layout in itself, excepted it is now possible to
create new force-based layouts more easily using BarnesHutLayout as a
base.

+++ src/org/graphstream/ui/layout/Layout.java
@@ -131,12 +131,12 @@ public interface Layout extends Pipe {
 	long getLastStepTime();
 
 	/**
-	 * The current layout algorithm quality. There are five quality levels. The
-	 * higher the level, the higher the quality.
+	 * The current layout algorithm quality. A number between 0 and 1
+	 * with 1 the highest (but probably slowest) quality.
 	 * 
-	 * @return A number between 0 and 4.
+	 * @return A number between 0 and 1.
 	 */
-	int getQuality();
+	double getQuality();
 
 	/**
 	 * The current layout force.
@@ -195,12 +195,13 @@ public interface Layout extends Pipe {
 	void setStabilizationLimit(double value);
 
 	/**
-	 * Set the overall quality level. There are five quality levels.
+	 * Set the overall quality level, a number between 0 and 1 with 1 the highest
+	 * quality available, but often with a slower computation.
 	 * 
 	 * @param qualityLevel
-	 *            The quality level in [0..4].
+	 *            The quality level, a number between 0 and 1.
 	 */
-	void setQuality(int qualityLevel);
+	void setQuality(double qualityLevel);
 
 	/**
 	 * If true, node informations messages are sent for every node. This is
